+++ node/src/types/node_config.rs
@@ -3,15 +3,37 @@ use serde::{Deserialize, Serialize};
 
 use crate::types::BlockHash;
 
+/// Maximum number of deploys to fetch in parallel, by default.
+const DEFAULT_MAX_PARALLEL_DEPLOY_FETCHES: u32 = 20;
+/// Maximum number of tries to fetch in parallel, by default.
+const DEFAULT_MAX_PARALLEL_TRIE_FETCHES: u32 = 20;
+
 /// Node fast-sync configuration.
-#[derive(Default, DataSize, Debug, Deserialize, Serialize, Clone)]
+#[derive(DataSize, Debug, Deserialize, Serialize, Clone)]
 // Disallow unknown fields to ensure config files and command-line overrides contain valid keys.
 #[serde(deny_unknown_fields)]
 pub struct NodeConfig {
     /// Hash used as a trust anchor when joining, if any.
     pub trusted_hash: Option<BlockHash>,
 
+    /// Maximum number of deploys to fetch in parallel.
+    pub max_parallel_deploy_fetches: u32,
+
+    /// Maximum number of trie nodes to fetch in parallel.
+    pub max_parallel_trie_fetches: u32,
+
     /// Whether to run in archival-sync mode. Archival-sync mode captures all data (blocks, deploys
     /// and global state) back to genesis.
     pub archival_sync: bool,
+}
+
+impl Default for NodeConfig {
+    fn default() -> NodeConfig {
+        NodeConfig {
+            trusted_hash: None,
+            max_parallel_deploy_fetches: DEFAULT_MAX_PARALLEL_DEPLOY_FETCHES,
+            max_parallel_trie_fetches: DEFAULT_MAX_PARALLEL_TRIE_FETCHES,
+            archival_sync: false,
+        }
+    }
 }